refactor: 검색 Controller, Service 제거
