+++ src/main/java/woowacrew/article/free/service/ArticleService.java
@@ -2,6 +2,7 @@ package woowacrew.article.free.service;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import woowacrew.article.free.domain.Article;
 import woowacrew.article.free.dto.ArticleRequestDto;
@@ -9,6 +10,7 @@ import woowacrew.article.free.dto.ArticleResponseDto;
 import woowacrew.article.free.dto.ArticleResponseDtos;
 import woowacrew.article.free.dto.ArticleUpdateDto;
 import woowacrew.article.free.utils.ArticleConverter;
+import woowacrew.search.domain.ArticleSearchSpec;
 import woowacrew.user.dto.UserContext;
 
 import java.util.List;
@@ -45,5 +47,13 @@ public class ArticleService {
 
     public void delete(Long articleId, UserContext userContext) {
         articleInternalService.delete(articleId, userContext);
+    }
+
+    public ArticleResponseDtos findSearchedArticles(String type, String content, Pageable pageable) {
+        Specification<Article> specification = ArticleSearchSpec.getSpec(type, content);
+        Page<Article> articlePages = articleInternalService.findAll(specification, pageable);
+        List<ArticleResponseDto> articleResponseDtos = ArticleConverter.articlePagesToArticleResponseDtos(articlePages);
+
+        return new ArticleResponseDtos(pageable.getPageNumber(), articlePages.getTotalPages(), articleResponseDtos);
     }
 }
