+++ src/test/java/woowacrew/article/free/service/ArticleServiceTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -18,6 +19,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -59,5 +61,27 @@ class ArticleServiceTest {
             articles.add(new Article(String.valueOf(i), String.valueOf(i), user));
         }
         return articles;
+    }
+
+    @Test
+    void 정상적으로_제목으로_게시글을_검색한다() {
+        String title = "hello";
+        String content = "bonjour";
+        User user = new User("userId", new Degree());
+
+        List<Article> articles = new ArrayList<>();
+        articles.add(new Article(title, content, user));
+        articles.add(new Article(title, content, user));
+        articles.add(new Article(title, content, user));
+
+        Page<Article> articlePages = new PageImpl<>(articles);
+
+        when(articleInternalService.findAll(any(), any())).thenReturn(articlePages);
+
+        ArticleResponseDtos actual = articleService.findSearchedArticles("title", "delete", PageRequest.of(0, 20));
+
+        assertThat(actual.getArticles().size()).isEqualTo(3);
+        assertThat(actual.getPageNumber()).isEqualTo(1);
+        assertThat(actual.getTotalPages()).isEqualTo(1);
     }
 }
